update position pid functions and add getspeed to raw motor

+++ src/BaseClass/BaseFeedbackMotor.h
@@ -11,11 +11,11 @@ public:
     virtual void update() = 0;
     virtual void reset() = 0;
 
-    virtual void setPositionTarget(float target) = 0;
+    virtual void setPositionTarget(long target) = 0;
     virtual void setVelocityTarget(float target) = 0;
 
     virtual bool isAtTarget() = 0;
-    virtual void setPositionTolerance(float tolerance) = 0;
+    virtual void setPositionTolerance(uint tolerance) = 0;
     virtual void setVelocityTolerance(float tolerance) = 0;
 
     virtual float getVelocity() = 0;
